+++ src/components/offers-empty/offers-empty.jsx
@@ -1,0 +1,20 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+export const OffersEmpty = ({selectedCity}) => (
+  <div className="cities">
+    <div className="cities__places-container cities__places-container--empty container">
+      <section className="cities__no-places">
+        <div className="cities__status-wrapper tabs__content">
+          <b className="cities__status">No places to stay available</b>
+          <p className="cities__status-description">We could not find any property availbale at the moment in {selectedCity}</p>
+        </div>
+      </section>
+      <div className="cities__right-section"></div>
+    </div>
+  </div>
+);
+
+OffersEmpty.propTypes = {
+  selectedCity: PropTypes.string.isRequired
+};
