+++ src/reducer/reducer.e2e.test.js
@@ -9,32 +9,47 @@ it(`Reducer without additional parameters should return initial state`, () => {
   });
 });
 
-it(`Reducer change city`, () => {
-  expect(reducer({
-    selectedCity: `Paris`,
-    currentOffer: null,
-    offers: DataValue.OFFERS_MOCK
-  },
-  ActionCreator.changeCity(`Hamburg`)
-  ))
-  .toEqual({
-    selectedCity: `Hamburg`,
-    currentOffer: null,
-    offers: DataValue.OFFERS_MOCK
+describe(`Reducer actions tests`, () => {
+  it(`Reducer change city`, () => {
+    expect(reducer({
+      selectedCity: `Paris`,
+      currentOffer: null,
+      offers: DataValue.OFFERS_MOCK
+    },
+    ActionCreator.changeCity(`Cologne`)
+    ))
+    .toEqual({
+      selectedCity: `Cologne`,
+      currentOffer: null,
+      offers: DataValue.OFFERS_MOCK
+    });
   });
-});
-
-it(`Reducer select offer`, () => {
-  expect(reducer({
-    city: `Paris`,
-    currentOffer: null,
-    offers: DataValue.OFFERS_MOCK
-  },
-  ActionCreator.selectOffer(DataValue.OFFERS_MOCK[2])
-  )).toEqual({
-    city: `Paris`,
-    currentOffer: DataValue.OFFERS_MOCK[2],
-    offers: DataValue.OFFERS_MOCK
+  it(`Reducer change getCityOffers`, () => {
+    expect(reducer({
+      selectedCity: `Paris`,
+      currentOffer: null,
+      offers: DataValue.OFFERS_MOCK
+    },
+    ActionCreator.getCityOffers(`Cologne`)
+    ))
+    .toEqual({
+      selectedCity: `Paris`,
+      currentOffer: null,
+      offers: []
+    });
+  });
+  it(`Reducer select offer`, () => {
+    expect(reducer({
+      city: `Cologne`,
+      currentOffer: null,
+      offers: DataValue.OFFERS_MOCK
+    },
+    ActionCreator.selectOffer(DataValue.OFFERS_MOCK[2])
+    )).toEqual({
+      city: `Cologne`,
+      currentOffer: DataValue.OFFERS_MOCK[2],
+      offers: DataValue.OFFERS_MOCK
+    });
   });
 });
